batch/pyproject.toml
[project]
name = "bds-collector"
version = "0.1.0"
description = "부동산 데이터 수집기 - 공공데이터포털 API를 사용하여 부동산 데이터를 수집하고 OpenSearch에 저장"
authors = [
    {name = "BDS Team", email = "team@bds.com"}
]
# readme = "README.md"  # Docker 빌드 시 문제 발생으로 주석 처리
requires-python = ">=3.11"
dependencies = [
    "requests",
    "pandas",
    "python-dotenv",
    "urllib3",
    "opensearch-py",
    "sentence-transformers",
    "scikit-learn",
    "numpy",
    "APScheduler",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
